requirements.txt
pyTelegramBotAPI==4.4.0
matplotlib==3.4.3
python-telegram-bot==20.3
sqlite3  
pydantic 
python-dotenv==1.0.0
google-auth==2.16.0
google-auth-oauthlib==0.5.3
google-api-python-client==2.62.0
